add (failing) unit tests (search dates in range)
In nvim-orgmore, I tried an agenda search (`<leader>oam`)
attempting to show only those items with a scheduled
date between two days ago and two days in the future:

> `Match: SCHEDULED>"<-2d>"&SCHEDULED<"<+2d>"`

somehow that didn't work, so I wrote this failing unit test to try and
find where the problem might be.  Note that searching for number ranges
(E.g., `PROP<6&PROP>3`) works as expected, it seems like the issue is
only with the date fields (`SCHEDULED`, `DEADLINE` - those these aren't
"properties" in the formal sense, but appear to be [treated as such][1]
for the purpose of making them searchable.

[1]: https://github.com/faizk/orgmode/blob/1d8c9b9417f8c8e9fb146d4f54fb1e90a4f7e534/lua/orgmode/files/file.lua#L283-L285